remove pkg/serviceaccount
