--- pkg/serviceaccount/OWNERS
@@ -1,13 +1,0 @@
-# See the OWNERS docs: https://git.k8s.io/community/contributors/guide/owners.md
-
-approvers:
-  - sig-api
-
-reviewers:
-  - sig-api
-
-labels:
-  - sig/api
-
-options:
-  no_parent_owners: true
